--- MakFood.Kitchen/MakFood.Kitchen.Domain/ConstituentAggrigate/Order/OrederState/CancelledOrderState.cs
@@ -1,10 +1,0 @@
-﻿namespace MakFood.Kitchen.Domain.Constituent.Order.OrederState
-{
-    public sealed class CancelledOrderState : OrderState
-    {
-        internal CancelledOrderState() { }
-        public override OrderStatus Status => OrderStatus.Cancelled;
-
-    }
-
-}
