--- MakFood.Kitchen/MakFood.Kitchen.Domain/ConstituentAggrigate/Order/Payment/PaymentMathods.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MakFood.Kitchen.Domain.ConstituentAggrigate.Order.Payment
-{
-    public enum PaymentMathods
-    {
-        Cash,
-        Wallet
-    }
-}
