--- MakFood.Kitchen/MakFood.Kitchen.Domain/ConstituentAggrigate/Order/Payment/PaymentState/PaymentStatus.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MakFood.Kitchen.Domain.ConstituentAggrigate.Order.Payment.PaymentState
-{
-    public enum  PaymentStatus
-    {
-
-    }
-}
