--- MakFood.Kitchen/MakFood.Kitchen.Domain/ConstituentAggrigate/Order/Status.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MakFood.Kitchen.Domain.Constituent.Order
-{
-    public enum Status
-    {
-        //Pending,
-        InvoiceSent,
-        Paid,
-        MiseOnPlace,
-        Cancelled
-
-    }
-}
