--- MakFood.Kitchen/MakFood.Kitchen.Domain/DiscountCode/DiscountPolicy/AllPermittedPolicy.cs
@@ -1,19 +1,0 @@
-﻿using MakFood.Kitchen.Domain.Base;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MakFood.Kitchen.Domain.DiscoudCode.DiscountPolicy
-{
-    public class AllPermittedPolicy :  DiscountPolicy 
-    {
-        public AllPermittedPolicy() 
-        {
-            var type = DiscountPolicyType.AllPermitted;
-        }
-        public override bool IsPermitted(Guid CamtomerId)=>true;
-    }
-}
-
