--- MakFood.Kitchen/MakFood.Kitchen.Domain/DiscountCode/Enum/DiscountPolicyType.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MakFood.Kitchen.Domain.DiscoudCode.DiscountPolicy
-{
-    public enum DiscountPolicyType 
-    {
-        AllPermitted,
-        SpecifiedPermision
-    }
-}
