+++ MakFood.Kitchen/MakFood.Kitchen.Domain/DiscountCodeAggrigate/DiscountPolicyAggrigate/AllPermittedPolicy.cs
@@ -1,0 +1,14 @@
+﻿using MakFood.Kitchen.Domain.DiscountCodeAggrigate.Enum;
+
+namespace MakFood.Kitchen.Domain.DiscountCodeAggrigate.DiscountPolicyAggrigate
+{
+    public class AllPermittedPolicy :  DiscountPolicy 
+    {
+        public AllPermittedPolicy() : base(DiscountPolicyType.AllPermitted)
+        {
+        }
+
+        public override bool IsPermitted(Guid CamtomerId)=>true;
+    }
+}
+
