+++ MakFood.Kitchen/MakFood.Kitchen.Domain/DiscountCodeAggrigate/Enum/DiscountPolicyType.cs
@@ -1,0 +1,8 @@
+﻿namespace MakFood.Kitchen.Domain.DiscountCodeAggrigate.Enum
+{
+    public enum DiscountPolicyType 
+    {
+        AllPermitted,
+        SpecifiedPermision
+    }
+}
