+++ MakFood.Kitchen/MakFood.Kitchen.Domain/MakFood.Kitchen.Domain.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Include="ConstituentAggrigate\Order\Payment\PaymentState\PaymentStatus.cs" />
+    <None Include="OrderAggrigate\OrderAggrigate\PaymentAggrigate\PaymentState\Enum\PaymentStatus.cs" />
   </ItemGroup>
 
   <ItemGroup>
