+++ MakFood.Kitchen/MakFood.Kitchen.Domain/OrderAggrigate/OrderAggrigate/Enum/OrderStatus.cs
@@ -1,0 +1,11 @@
+﻿namespace MakFood.Kitchen.Domain.OrderAggrigate.OrderAggrigate.Enum
+{
+    public enum OrderStatus
+    { 
+        InvoiceSent,
+        Paid,
+        MiseOnPlace,
+        Cancelled
+
+    }
+}
