+++ MakFood.Kitchen/MakFood.Kitchen.Domain/OrderAggrigate/OrderAggrigate/OrederState/CancelledOrderState.cs
@@ -1,0 +1,10 @@
+﻿namespace MakFood.Kitchen.Domain.OrderAggrigate.OrderAggrigate.OrederState
+{
+    public sealed class CancelledOrderState : OrderState
+    {
+        internal CancelledOrderState() { }
+        public override OrderStatus Status => OrderStatus.Cancelled;
+
+    }
+
+}
