+++ MakFood.Kitchen/MakFood.Kitchen.Domain/OrderAggrigate/OrderAggrigate/PaymentAggrigate/PaymentState/Enum/PaymentStatus.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MakFood.Kitchen.Domain.OrderAggrigate.OrderAggrigate.PaymentAggrigate.PaymentState.Enum
+{
+    public enum  PaymentStatus
+    {
+        Aproved,
+        Cancelled,
+        Paied
+    }
+}
